Trigger Rework
I changed trigger functionality to allow for more stuff like Intervals for example. Should work.

Sadly noticed another bigger oversight:
Currently a trigger just calls ALL Effects registered to it (And probably on all Sources as well? or not because sources are in the coroutines as well? dunno). I somehow need to differentiate wich gets triggered when.