TestFramework V1
Setup the first iteration of test framework.
Fixed some Init timing problems with triggers and some null refs in the Registry check.
Changed how Stats work to make them more fexible.

To register trigger for late init, just add their Init Method to the Event in the ItemEventhandler.

For now doubling the same effect on an item does nothing...need to find an elegant way to handle that.