Class/Type structure fix
Fixed the structural problem i found earlier, by replacing the singular type in the class with a list of allowed types and adding a selection wich type to use to the item.

I also updated the editor script to show a dropdown menu to select the used type inside the class instead of a integer field inside the item.