UI Inspector Panel
Added an inspector panel to the tool, allowing to see and edit exposed properties of the SOs.

Added a custom tag to mark allowed properties that should be editable in the tool.

Rearranged some conent into new scripts to improve readability.

ToDo:
Add the missing types/dictionaries etc. to the inspector panel.

Add functionality that changes property content when changed in editor tool.