Sub Modules
Added a new layer of tabs allowing for switching from item modules to the other sub modules like item containers or stats.

Added Sub modules to the editor.

Added basic Settings SO to store settings as well as a settings tab to alter the data.

ToDo:
CLean up code and try to replace all literal string paths etc. with dynamic variants.

ToDo:
Check if dynamic version of Stat Value can be added to Abstract SO script for toolkit access.

ToDo:
Make it so the creator window type selection is automatically for tab-type/ treeview selected type?

ToDo:
Module to Json and Json to module.