Basic File save and load
Added basic Module to JSON and JSON to module. Structure needs to be improved and usability of UI needs to be improved as well, but for now its there.

ToDo:
Check if dynamic version of Stat Value can be added to Abstract SO script for toolkit access.

ToDo:
Make it so the creator window type selection is automatically for tab-type/ treeview selected type?